requirements.txt
flask
supabase
python-dotenv
requests
selenium
webdriver-manager
google-generativeai
statistics